+++ src/test/kotlin/com/kh/api/KakaoSkillApiTest.kt
@@ -2,13 +2,11 @@ package com.kh.api
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.kh.api.config.AppConfig
-import com.kh.api.request.SearchTrainParams
 import com.kh.api.request.SkillPayload
 import com.kh.occupying.Korail
-import com.kh.occupying.dto.response.SearchResponse
 import com.kh.service.BackgroundExecutor
+import com.kh.util.RequestBodyCreator
 import com.kh.util.SecretProperties
-import com.kh.util.mapTo
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -41,6 +39,8 @@ class KakaoSkillApiTest {
     @MockBean
     private lateinit var backgroundExecutor: BackgroundExecutor
 
+    private lateinit var requestBodyCreator: RequestBodyCreator
+
     private lateinit var id: String
     private lateinit var pw: String
     private lateinit var departureDate: String
@@ -69,12 +69,14 @@ class KakaoSkillApiTest {
 
         departureDate = LocalDate.now().plusDays(1)
                 .format(DateTimeFormatter.ofPattern("yyyyMMdd"))
+
+        requestBodyCreator = RequestBodyCreator(korail)
     }
 
     @Test
     fun `test find trains`() {
         // Arrange
-        val body = findTrainRequest(
+        val body = requestBodyCreator.findTrainRequest(
                 departureDate = departureDate,
                 departureTime = "070000",
                 departureStation = "서울",
@@ -104,7 +106,7 @@ class KakaoSkillApiTest {
             expected: String
     ) {
         // Arrange
-        val body = findTrainRequest(
+        val body = requestBodyCreator.findTrainRequest(
                 departureDate = departureDate,
                 departureTime = departureTime,
                 departureStation = departureStation,
@@ -130,7 +132,7 @@ class KakaoSkillApiTest {
     @Test
     fun `test find trains response failed`() {
         // Arrange
-        val body = findTrainRequest(
+        val body = requestBodyCreator.findTrainRequest(
                 departureDate = "20180101",
                 departureTime = "070000",
                 departureStation = "서울",
@@ -153,7 +155,15 @@ class KakaoSkillApiTest {
     @Test
     fun `test reserve train`() {
         // Arrange
-        val body = reservationRequest(id, pw, email)
+        val body = requestBodyCreator.reservationRequest(
+                departureDate = departureDate,
+                departureTime = "070000",
+                departureStation = "서울",
+                destinationStation = "부산",
+                id = id,
+                pw = pw,
+                email = email
+        )
 
         // Act & Assert
         webClient.post()
@@ -171,119 +181,4 @@ class KakaoSkillApiTest {
         BDDMockito.verify(backgroundExecutor).reserveTrain(payload)
     }
 
-    private fun findTrainRequest(
-            departureDate: String,
-            departureTime: String,
-            departureStation: String,
-            destinationStation: String
-    ): String {
-        return """
-                {
-                  "intent": {
-                    "id": "d7coyklcoly0q5lhjh6zdjzd",
-                    "name": "블록 이름"
-                  },
-                  "userRequest": {
-                    "timezone": "Asia/Seoul",
-                    "params": {
-                      "ignoreMe": "true"
-                    },
-                    "block": {
-                      "id": "d7coyklcoly0q5lhjh6zdjzd",
-                      "name": "블록 이름"
-                    },
-                    "utterance": "발화 내용",
-                    "lang": null,
-                    "user": {
-                      "id": "515520",
-                      "type": "accountId",
-                      "properties": {}
-                    }
-                  },
-                  "bot": {
-                    "id": "5d97076fb617ea00012af4e0",
-                    "name": "봇 이름"
-                  },
-                  "action": {
-                    "name": "g0xxzoae15",
-                    "clientExtra": null,
-                    "params": {
-                      "departure-date": "$departureDate",
-                      "departure-time": "$departureTime",
-                      "departure-station": "$departureStation",
-                      "destination-station": "$destinationStation"
-                    },
-                    "id": "krcu052b90c6angense4fxgy"
-                  }
-                }
-            """.trimIndent()
-    }
-
-    private fun reservationRequest(id: String, pw: String, email: String): String {
-        val trainNo = getTrainNo()
-        return """
-                {
-                    "intent": {
-                    "id": "d7coyklcoly0q5lhjh6zdjzd",
-                    "name": "블록 이름"
-                  },
-                  "userRequest": {
-                    "timezone": "Asia/Seoul",
-                    "params": {
-                      "ignoreMe": "true"
-                    },
-                    "block": {
-                      "id": "d7coyklcoly0q5lhjh6zdjzd",
-                      "name": "블록 이름"
-                    },
-                    "utterance": "발화 내용",
-                    "lang": null,
-                    "user": {
-                      "id": "515520",
-                      "type": "accountId",
-                      "properties": {}
-                    }
-                  },
-                  "bot": {
-                    "id": "5d97076fb617ea00012af4e0",
-                    "name": "봇 이름"
-                  },
-                    "action": {
-                        "name": "예약 API",
-                        "clientExtra": {
-                            "departure-time": "070000",
-                            "departure-date": "$departureDate",
-                            "train-no": "$trainNo",
-                            "destination-station": "부산",
-                            "departure-station": "서울"
-                        },
-                        "params": {
-                            "id": "$id",
-                            "pw": "$pw",
-                            "email": "$email"
-                        },
-                        "id": "5d9dc807ffa7480001dace7e"
-                    }
-                }
-            """.trimIndent()
-    }
-
-    private fun getTrainNo(): String {
-        val request = jacksonObjectMapper()
-                .readValue(findTrainRequest(
-                        departureDate = departureDate,
-                        departureTime = "070000",
-                        departureStation = "서울",
-                        destinationStation = "부산"
-                ), SkillPayload::class.java)
-        val payload = request.action.params
-                .mapTo<SearchTrainParams>()
-                .getSearchParams()
-        return korail.search(payload).map {
-            (it as SearchResponse).train.items
-                    .map { response -> response.toDomain() }
-                    .first { train -> train.hasSeat() }
-                    .no
-        }.block()!!
-    }
 }
